session管理をhook化
